State machine
split up lobby state machine and dodgeball state machine
